gem test deps: update selenium, rack. Rm httponly rack-test driver test
rack-test no longer implements httponly (according to gemini - not verified).
It's easier to remove this test than to support before and after separately.
It's not part of the api anyway, and people would have already seen this
breaking their tests if they were using it for that.
If the gem was going forward it would be a major version breaking change
to remove httponly from the rack-test cookies, or always return it as false.
Instead, people will always see a false value with rack-test once they
update rack-test.